fix(semantics): extend cross-validation to DEFINE_PACKET_ID, packet_headers enum, and zc_par_4job
Items 1-4 from the unverifiable audit:

Item 1: Add DEFINE_PACKET_ID to scanner (4 more verified bindings)
  crossvalidate_test.go: definePacketHeaderRe now matches both
  DEFINE_PACKET_HEADER and DEFINE_PACKET_ID patterns.

Item 2: self_chat:0x008e → PACKET_ZC_NOTIFY_PLAYERCHAT confirmed correct
  clif.cpp:6661 comment: '008e <packet len>.W <message>.?B (ZC_NOTIFY_PLAYERCHAT)'
  Added to KnownCorrectBindings.

Item 3+4: Add packet_headers enum parser (19 old-style structs now verified)
  parsePacketHeadersEnum() reads ALL PACKETVER branches of the enum and
  accumulates all historical IDs per type. structSuffix() maps
  'packet_idle_unit' → 'PACKET_IDLE_UNIT' (the enum key convention).
  Result: verified count 386 → 408, not_found 26 → 8.

New mismatch caught and fixed:
  stat_update: 0x0B25 had wrong struct PACKET_ZC_LONGLONGPAR_CHANGE
  DEFINE_PACKET_ID says 0x0B25 = ZC_PAR_4JOB_CHANGE (4th job stat)
               and     0x0ACB = ZC_LONGLONGPAR_CHANGE (base stat u64)
  Fixed in DB via MCP; decode/events files regenerated.

Remaining 8 not-found: all verified manually via clif.cpp/clif_packetdb.hpp
  and documented in KnownCorrectBindings with source citations.

+++ pkg/events/stat_update.go
@@ -6,7 +6,7 @@ package events
 type StatUpdate struct {
 	VarID uint16
 	Count int32
-	Amount int64
+	Amount int32
 	StatusID uint16
 	Value uint8
 }
